Take out --family-for-fp option which is not entirely correct (#6329).
This option looks for relays having the given fingerprint in their
family line, but it doesn't cross-check if the given relay has those
relays in its family line, too.  The option also doesn't look at
nicknames (of relays with the Named flag) in family lines, which means
it might miss some family members.