Fix the way how advertised bandwidth is supposed to be used (#6232).
A relay's advertised bandwidth is not supposed to be used as replacement
for a potentially missing consensus bandwidth.  The idea is to compare
path selection probability based on reported vs measured bandwidth.  There
should be an entropy (and max entropy) based on advertised bandwidths in
addition to the same numbers for consensus weights.

Also removing the hex_digest field, because we only need it once when
looking up the advertised bandwidth.

Oh, and changing the default advertised bandwidth from None to 0, so that
we can calculate with the result.  There may always be single server
descriptors missing.