Treat unresolved AS numbers correctly (#6232).
So far, relays that couldn't be found in the AS database were ignored in
the AS entropy calculation.  That's not correct.  We have to treat them as
"other" AS, similar to how we handle relays for which we couldn't find the
country code.  Otherwise the probability of all ASes wouldn't sum up to 1.