Add 'no more than 2 per /24' clause
Only 2 relays per /24 blocks are used. Figure out the network
addr for a relay and store it in a dict(network_data) with a
'network'->['relay', 'relay'] mapping. Check if 2 relays are
present in network_data for a particular network, if yes, then
find the relay with the smallest exit probability and remove
it from the list of all relays and add the current relay to
the list of all relays.